Revert "Merge pull request #359 from dbt-labs/migrate-webhook-to-framework"
This reverts commit 7bb07afdeed091a81493416038fc5fdc5a35864f, reversing
changes made to 14ac8cb001712777323946f22974676d8e9fbc85.

+++ pkg/dbt_cloud/webhook.go
@@ -32,7 +32,7 @@ type WebhookWrite struct {
 	Description string   `json:"description,omitempty"`
 	ClientUrl   string   `json:"client_url"`
 	EventTypes  []string `json:"event_types,omitempty"`
-	JobIds      []int64  `json:"job_ids"`
+	JobIds      []int    `json:"job_ids"`
 	Active      bool     `json:"active,omitempty"`
 }
 
@@ -71,7 +71,7 @@ func (c *Client) CreateWebhook(
 	description string,
 	clientUrl string,
 	eventTypes []string,
-	jobIds []int64,
+	jobIds []int,
 	active bool,
 ) (*WebhookRead, error) {
